+++ tensorflow/lite/toco/graph_transformations/remove_successive_transpose.cc
@@ -31,12 +31,12 @@ bool TransformsToIdentity(std::vector<int> const& perm1,
   // perm1 is the order of the indices after first transpose. When perm1 is
   // reordered according to perm2, if the result is simple increasing sequence
   // i.e., range(0, perm1.size()), then the two transposes cancel each other.
-  for (int i = 0; i < perm1.size(); ++i) {
-    if (perm1[i] < 0 || perm1[i] >= perm1.size() || perm2[i] < 0 ||
-        perm2[i] >= perm1.size()) {
+  for (size_t i = 0; i < perm1.size(); ++i) {
+    if (perm1[i] < 0 || perm1[i] >= static_cast<int>(perm1.size()) || perm2[i] < 0 ||
+        perm2[i] >= static_cast<int>(perm1.size())) {
       return false;
     }
-    if (perm1[perm2[i]] != i) {
+    if (perm1[perm2[i]] != static_cast<int>(i)) {
       return false;
     }
   }
@@ -46,7 +46,7 @@ bool TransformsToIdentity(std::vector<int> const& perm1,
 void ReplaceOpInputsWith(Model* model, const string& lookfor,
                          const string& replacewith) {
   for (const auto& op : model->operators) {
-    for (int i = 0; i < op->inputs.size(); ++i) {
+    for (size_t i = 0; i < op->inputs.size(); ++i) {
       if (op->inputs[i] == lookfor) {
         op->inputs[i] = replacewith;
       }
